backend/requirements.txt
requests
beautifulsoup4
psycopg2-binary
Flask==2.3.2
Flask-Cors==1.10.3
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
